+++ resources/views/admin/pages/index.blade.php
@@ -68,8 +68,12 @@
                                       Action
                                       </button>
                                       <div class="dropdown-menu arrow " id="options">
-                                          <a class="dropdown-item" href="{{ route('admin.page.edit',$pageInfo->id) }}"><i class="ft-edit green"></i> Edit </a>
-                                          <a data-id="{{$pageInfo->id}}" onclick="deleteFunction({{$pageInfo->id}})"  id="delete" data-toggle="modal" data-backdrop="false" data-target="#info" class="dropdown-item" href="#"><i class="ft-slash red"></i> Delete</a>
+                                          @can('edit', new App\Models\Page)
+                                            <a class="dropdown-item" href="{{ route('admin.page.edit',$pageInfo->id) }}"><i class="ft-edit green"></i> Edit </a>
+                                          @endcan
+                                          @can('delete', new App\Models\Page)
+                                            <a data-id="{{$pageInfo->id}}" onclick="deleteFunction({{$pageInfo->id}})"  id="delete" data-toggle="modal" data-backdrop="false" data-target="#info" class="dropdown-item" href="#"><i class="ft-slash red"></i> Delete</a>
+                                          @endcan
                                       </div>
                                   </div>
                               </div>
